+++ app/index.tsx
@@ -2,7 +2,7 @@ import { useRef, useState } from 'react';
 import { View, Text, ScrollView } from 'react-native';
 import WebView, { WebViewMessageEvent } from 'react-native-webview';
 import { config } from '@/src/utils/domainConfig';
-import { handleWebViewMessage } from '@/src/utils/webviewBridge';
+import { handleWebViewMessage } from '@/src/utils/clipBaordwebviewBridge';
 
 export default function Index() {
   const webViewRef = useRef<WebView>(null);
