+++ src/constants/clipboardConst.ts
