+++ src/types/clipboardTypes.ts
